Factor out DeleteConfirmation and NotificationPopup
