+++ client/app/bundles/course/survey/actions/index.js
@@ -1,24 +1,3 @@
-import actionTypes from '../constants';
+import { setNotification, resetDeleteConfirmation, showDeleteConfirmation } from 'lib/actions';
 
-export function resetDeleteConfirmation() {
-  return { type: actionTypes.RESET_DELETE_CONFIRMATION };
-}
-
-export function showDeleteConfirmation(onConfirm) {
-  return (dispatch) => {
-    const confirmAndDismiss = () => {
-      onConfirm();
-      dispatch(resetDeleteConfirmation());
-    };
-    dispatch({ type: actionTypes.SHOW_DELETE_CONFIRMATION, onConfirm: confirmAndDismiss });
-  };
-}
-
-export function setNotification(message) {
-  return (dispatch) => {
-    dispatch({
-      type: actionTypes.SET_SURVEY_NOTIFICATION,
-      message,
-    });
-  };
-}
+export { setNotification, resetDeleteConfirmation, showDeleteConfirmation };
