+++ client/app/bundles/course/survey/constants.js
@@ -78,9 +78,6 @@ const actionTypes = mirrorCreator([
   'QUESTION_FORM_HIDE',
   'SECTION_FORM_SHOW',
   'SECTION_FORM_HIDE',
-  'SHOW_DELETE_CONFIRMATION',
-  'RESET_DELETE_CONFIRMATION',
-  'SET_SURVEY_NOTIFICATION',
   'SET_DRAGGED_QUESTION',
   'REORDER_QUESTION',
   'CHANGE_QUESTION_SECTION',
