+++ client/app/bundles/course/survey/containers/SurveysLayout/index.jsx
@@ -1,15 +1,14 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import { withRouter, Route, Switch } from 'react-router-dom';
-import NotificationBar, { notificationShape } from 'lib/components/NotificationBar';
+import NotificationPopup from 'lib/containers/NotificationPopup';
+import DeleteConfirmation from 'lib/containers/DeleteConfirmation';
 import SurveyFormDialogue from 'course/survey/containers/SurveyFormDialogue';
 import QuestionFormDialogue from 'course/survey/containers/QuestionFormDialogue';
 import SectionFormDialogue from 'course/survey/containers/SectionFormDialogue';
 import SurveyLayout from 'course/survey/containers/SurveyLayout';
 import SurveyIndex from 'course/survey/pages/SurveyIndex';
-import DeleteConfirmation from './DeleteConfirmation';
 
-const SurveysLayout = ({ notification }) => {
+const SurveysLayout = () => {
   const surveyRoot = '/courses/:courseId/surveys';
   return (
     <div>
@@ -17,7 +16,7 @@ const SurveysLayout = ({ notification }) => {
       <QuestionFormDialogue />
       <SectionFormDialogue />
       <DeleteConfirmation />
-      <NotificationBar notification={notification} />
+      <NotificationPopup />
 
       <Switch>
         <Route exact path={surveyRoot} component={SurveyIndex} />
@@ -27,8 +26,4 @@ const SurveysLayout = ({ notification }) => {
   );
 };
 
-SurveysLayout.propTypes = {
-  notification: notificationShape,
-};
-
-export default withRouter(connect(({ notification }) => ({ notification }))(SurveysLayout));
+export default withRouter(SurveysLayout);
