+++ client/app/bundles/course/survey/pages/SurveyShow/Section/__test__/DeleteSectionButton.test.jsx
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom';
 import { mount } from 'enzyme';
 import ReactTestUtils from 'react-dom/test-utils';
 import CourseAPI from 'api/course';
+import DeleteConfirmation from 'lib/containers/DeleteConfirmation';
 import storeCreator from 'course/survey/store';
-import DeleteConfirmation from 'course/survey/containers/SurveysLayout/DeleteConfirmation';
 import DeleteSectionButton from '../DeleteSectionButton';
 
 describe('<DeleteSectionButton />', () => {
