+++ client/app/bundles/course/survey/reducers/index.js
@@ -1,18 +1,20 @@
 import { combineReducers } from 'redux';
 import { reducer as formReducer } from 'redux-form';
+import notificationPopup from 'lib/reducers/notificationPopup';
+import deleteConfirmation from 'lib/reducers/deleteConfirmation';
 import surveys from './surveys';
 import surveyForm from './surveyForm';
 import questionForm from './questionForm';
 import responseForm from './responseForm';
 import sectionForm from './sectionForm';
-import deleteConfirmation from './deleteConfirmation';
-import notification from './notification';
 import surveysFlags from './surveysFlags';
 import results from './results';
 import responses from './responses';
 
 
 export default combineReducers({
+  notificationPopup,
+  deleteConfirmation,
   surveys,
   responses,
   results,
@@ -20,8 +22,6 @@ export default combineReducers({
   questionForm,
   responseForm,
   sectionForm,
-  deleteConfirmation,
-  notification,
   surveysFlags,
   form: formReducer,
 });
