+++ client/app/lib/constants/index.js
@@ -1,0 +1,9 @@
+import mirrorCreator from 'mirror-creator';
+
+const actionTypes = mirrorCreator([
+  'SET_NOTIFICATION',
+  'SHOW_DELETE_CONFIRMATION',
+  'RESET_DELETE_CONFIRMATION',
+]);
+
+export default actionTypes;
