+++ client/app/lib/reducers/notificationPopup.js
